requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
drf-yasg==1.20.0
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
motor==2.5.1
packaging==21.3
psycopg2==2.7.7
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.16.0
zipp==3.6.0
